Fix attribute passing to the mirror deleter
